Precisionpulse1/backend/requirements.txt
flask
flask-socketio
eventlet
paho-mqtt
